factor out Ohua components
Components rewritten for Ohua
compilation should get their own
modules, existing in parallel
to the original ones. This
way we can have only supported
code in the Ohua modules to
improve clarity of the code as well
as of the rewrite process itself.

This commit starts separating the
Ohua versions for Interface and tcp socket.