adapted, no UDP support in Ohua version
There is now a separate ohua version of the interface to
a) to have both versions to compare and b) to be able to
cut out some, currently not supported features to make it
less confusing.

The downside is, that a separate interface requires a separate
inner interface which is argument to functions of the sockets.
This is currently only implemented for tcp. A possible temporary
remedy might be to make the inner interface public.

We (currently) do not change it and this way I can support all sockets
and keep the other interface tests running also for the Ohua interface