encapsulate app, 'sending' sockets
We need a way for the app code to
interact with sockets and make those interactions
effective in the sockets used by the tcp_ip_stack

As the SocketSet is currently not a part of the
interface any more, the most obvious way is to
use the SocketSet itself as 'message format'.

The problem is that, that we currently can not,
and maybe after all don't want to serialize the
SocketSet, which we needed to do to make this a
distributed program.