outer loop as recursion, sockets as object
Approaching a compilable state, the outer
application loop of the Ohua-k-v-store is now a
recursion. Further sockets and device are now passed as
actual objects, not as reference any more.

Finally the match expression has been moved out of compile scope.
Learnings and rationals are described in the according commit in my
'sketch repo'
https://github.com/Pagolin/sketch_transformations/commit/951a62879812b0ac0feb64818060efdb1430255b