current state of poll_egress
The poll_egress_4_ask (as the name might suggest)
is the current rewrite stage of the poll function
concentrating on egress first. The goal is to make poll compile
without passing any of our target states as argument.

Changes to device and implementation of LocalToken
enable a decoupling of packet generation from packet sending.

The current state does not compile. This is on purpose, because
there might be different ways to satisfy the borrow checker.
Making changes to the structs (in particular interface) is one of them
but I'm not sure if and how many of the borrow problem still
exists after proper rewrite and encapsulation. Therefor the
current version is also in the original interface, not in the
already adapted interface Ohua. I want to see and discuss the
problems along the way.