recursion in poll, tests fixed
I started approaching the problem form a second
direction, namely restructuring poll (or clones of poll build along the way)
in place to be able to test it continuously.

We'll see if the approaches merge later.

Also at least basic tests for egress Ohua interface vs
original interface and for the current refactoring state of
poll work again :-)