replaced Token passing
The InnerInterface was given a reference to the device (the sending token) to call inside
the interface code. Obviously that won't work in our target program.

So now, part of the device state function, namely the function of the
token to provide a method call
`token.consume(f:/*closure writing packet to device buffer*/)`
is replaced by a `LocalToken`. This keeps the InnerInterface code stable
while we are able to extract the packet and pass is to the device in scope.

On the device side this meant one actually required change and one that
is just needed to keep code compiling while we work on it.

Change 1 is to write a wrapper `device.consume()` because this is an
operation on the device state and that needs to be obvious to the compiler.

Change 2 is a temporary replacement of the `device.transmit()` and new
`device.cosume()` methods. Just as later in Ohua, Rust wont let us mutably
borrow the device twice in a loop. So while we need to fix this 'double
device use' anyways, we enable keeping it this way to be able to test our
next modifications