for + break -> while; dummy methods
We did 2 things here. First we transformed
the for loop with break statement to a while loop.
While loops are closer to recursion (actually just syntax sugar for it)
but currently have the advantage that we can keep the code inside the method
instead of having to use yet another method. Which is for know better to keep
the overview.

Second: Knowing that the inner interface doesn't need to be mutable for our basic
tests we also introduced interim replacements for the interface methods that only
borrow immutably and use a mock InterfaceInner to feed it to lower level methods.
This way we can keep basic tests running while refactoring to a state that can again use the
real methods. (Compiler will hopefully later not need to run tests to see if he's
transforming code correctly :-D)