egress state inside interface
Communication between the interface and the device need not/should not include
sockets. We could figure that out via a dependency analysis, leading to the conclusion
that interface should take the sockets and keep them, and the current state
of processing egress, while only communicating messages to the device.

This is reflected in this attempt to keep the 'EgressState' inside the interface.
It would be easier (in terms of managing lifetimes and references) to keep the
information about the currently used socket as either an actually moved socket item or
just a socket handle. However I tried to keep the original data structures and
handle the socket iterator and a '&mut Item' to ease argumentation about automatic
transformations later.

Does not compile due to lifetime and borrow issues.