wip: enums for IfaceCall, DevCall, CHAOS
To make our 'stateful components' (interface and device) composeable
they essentially need to have one 'entry function'.
This function should take a 'call' argument including the required
arguments and dispatch internally to the required function.
It should return a call to another node basically. In case of the device,
it should always return an IfaceCall, in case of the interface the
'entry function' should return an Either<DevCall, AppCall>.

Currently everything is chaos though i.e. nothing compiles and there
are a bunch of methods that appeared during refactoring/encapsulation
that need to be deleted later.