move 'poll_egress' into scope
So we're on the way to have only one stateful call for iface and device in
poll (at least while we only look at egress). Now we moved this into scope and actually
threaded (moved) the device, interface and sockets through instead of using refs.

Still borrowing and lifetimes are broken, mainly because handling sockets and the current state
of iteration in the interface is not working as is.

So next step is merging the 'egress call chain' in scope and finding a solution for the keeping the iteration state in the interface (Taht is hopeful as close as possible to something a compiler could come up with itself.)