moved poll-loop into iface.process_call, still BROKEN
The interface needs to control whether to stay in one of the inner loops
(poll, egress, ingress) or return to the outer loop of the 'composition'.
So the ability to call another poll loop on itself is now build into the
process_call, eliminating one (compile scope) visible recursion.

There are still lifetime and borrowing problems around keeping and using the
'EgressState' in the interface.