owned IpPacket - compiles
The problem was, similar to the one wwith the iterator state, that we can not have two references to the same
memory saved in the state. But this was exactly what happend, because the TcpRepr contained a buffer, pointing
to the sockets storage. So when we produced a Packets from a Socket we had two borrows on the SocketSet.
There are/where 3 options to handle this
1. Make all representations of sendable stuff owned -> complicated, hardly doable for a compiler
2. Produce packet several times. At first to check whether there can be a packet, then to send it, then to
   get it again for post processing. Doing so we can drop the extra reference to the Socketset at each
   function return. This actually reflects the general time/memory problem we have with distribution.
   (Either you send and store -> more memory, or you calculate twice -> more time/cycles)
3. The option we've chosen for now. Write an owned version of the packet representation to store data and
   convert back to reference (this time to the owned version, not to the sockets), when the packet is used