sending only messages - stack overflow
Sending the SocketSeet between App and Interface caused the interesting error of
'state leaking into function call': Basically the problem is, that SocketSets contain Sockets with lifetimed
buffers, so we can not 'send them', since the compiler can not guarantee that the sockets we give the interface will be the same that we get back and that either will persist through calls as their lifetime comes from outer scope.

So we had (again) the options to make them owned, or to make them local to one components and tie their lifetime to this component. We chose the second option. SO now the sockets are part of the interface again, as the interface makes more elaborate use of their implementation. Tha App only processes a vector of messages per SocketHandle.
To do so we also had to move the remaining socket handling back to the interface.
Its now split in two states, namely 'loading' the sockets with answers from the app before poll and 'unloading' socket
answers after the poll call.