runs but logic problem
We can not use the recursion we would feed to Ohua, because Rust does not by default (and surely not in thi case)
support tail recursion so we get a stack overflow using recursion.

Thats whhy for now we use a loop again. The current state runs but fails with 'No route to host' so there is obviously a problem in the logic still.