dummy replacement for phy_wait
We needed to replace `phy_wait`, because waiting for a file descriptor
will not work this way in m3.

We will need to use one of the devices m3 provides. They implement a
method to ask whether they should be polled, which we mimic here.

Another constraint is, that we need to 'wait' in the main scope. This is because
we will need a handle to the running action and we don't want to mix the code for
the runtime (i.e. getting a handle on the running action), with the 'pure' smoltcp
code. Finally we need to have that wait between a device call and an interface call.

The reason is, that it needs to happen after the socket(s) received its/there answer(s).
So we can not do it while we call the app. (Well, maybe we could but lets stick with
the original order of states as far as possible).

So now we 1. ask the Interface for `poll_delay`, send a request to the device to add
it's (dummy) `needs_poll` and send those two back into scope to use it in a dummy waiting function.