using phy_wait again under the hood
As we do not use the original smoltcp devices anyway in m3 it's ok
to use `phy_wait` outside the compile scope. So now we do this and thereby
reduce CPU costs of running the app significantly.

Also fixed the return type of `device.process_call`, returning either a
call to the interface or a signal to wait now. Thereby we eliminate the inconsistent
`InterfaceCall::ProcessWait`, that was never ment to be processed by the interface.