double recursion
So we had a nice loop. But Ohua can not process loops.

So we refactored to a simple recursion.But we still had the problem, that device and app are called in if-else-branches. However Ohua currently does not support stateful calls in branches.

So we refactored to two recursions. But still we had the problem of having to unwrap the
Either<DeviceCall, AppCall> inside the recursion branches, because only there we know what we'll
get.

Still we have an unwrap left. So next we need to just thread through the Options and unwrap somewhere out of scope.